[logging] WIP > First pass on extracting the logging logic from "estransport".Client.Perform()
